LayoutTests/editing/selection/ios/move-by-word-with-keyboard.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<html>
<head>
<meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1, shrink-to-fit=no">
<script src="../../../resources/ui-helper.js"></script>
<script src="../../../resources/dump-as-markup.js"></script>
<script>

Markup.description(`This tests moving the caret with word granularity on iOS.
WebKit should stop at the end of each word before whitespace when moving forwards, and after whitespace when moving backwards,
The empty line should be skipped in either direction.`);
Markup.noAutoDump();
Markup.waitUntilDone();

async function runTest()
{
    const editor = document.getElementById('editor');

    if (window.testRunner)
        await UIHelper.activateElementAndWaitForInputSession(editor);

    if (!window.testRunner) {
        Markup.dump(editor, 'Markup');
        Markup.notifyDone();
        return;
    }

    getSelection().setPosition(editor, 0);
    Markup.dump(editor, 'Before moving forwards');

    const moveCount = 6;

    for (let i = 0; i < moveCount; i++) {
        await UIHelper.keyDown('rightArrow', ['altKey']);
        await UIHelper.ensurePresentationUpdate();
        Markup.dump(editor, `After moving to the right by word (${i + 1})`);
    }

    getSelection().setPosition(editor, editor.childNodes.length);
    Markup.dump(editor, 'Before moving backwards');

    for (let i = 0; i < moveCount; i++) {
        await UIHelper.keyDown('leftArrow', ['altKey']);
        await UIHelper.ensurePresentationUpdate();
        Markup.dump(editor, `After moving to the left by word (${i + 1})`);
    }

    getSelection().setPosition(editor, 0);
    Markup.dump(editor, 'Before extending forwards');

    for (let i = 0; i < moveCount; i++) {
        await UIHelper.keyDown('rightArrow', ['shiftKey', 'altKey']);
        await UIHelper.ensurePresentationUpdate();
        Markup.dump(editor, `After extending to the right by word (${i + 1})`);
    }

    getSelection().setPosition(editor, editor.childNodes.length);
    Markup.dump(editor, 'Before extending backwards');

    for (let i = 0; i < moveCount; i++) {
        await UIHelper.keyDown('leftArrow', ['shiftKey', 'altKey']);
        await UIHelper.ensurePresentationUpdate();
        Markup.dump(editor, `After extending to the left by word (${i + 1})`);
    }

    Markup.notifyDone();
}

onload = runTest;
</script>
</head>
<body>
<div id="editor" contenteditable>
    <div>hello world</div>
    <br>
    <div>red, green, & blue.</div>
</div>
</body>
</html>
