LayoutTests/editing/selection/ios/move-by-word-with-keyboard-expected.txt

This tests moving the caret with word granularity on iOS.
WebKit should stop at the end of each word before whitespace when moving forwards, and after whitespace when moving backwards,
The empty line should be skipped in either direction.

Before moving forwards:
| "
    "
| <div>
|   "<#selection-caret>hello world"
| "
    "
| <br>
| "
    "
| <div>
|   "red, green, & blue."
| "
"

After moving to the right by word (1):
| "
    "
| <div>
|   "hello<#selection-caret> world"
| "
    "
| <br>
| "
    "
| <div>
|   "red, green, & blue."
| "
"

After moving to the right by word (2):
| "
    "
| <div>
|   "hello world<#selection-caret>"
| "
    "
| <br>
| "
    "
| <div>
|   "red, green, & blue."
| "
"

After moving to the right by word (3):
| "
    "
| <div>
|   "hello world"
| "
    "
| <br>
| "
    "
| <div>
|   "red,<#selection-caret> green, & blue."
| "
"

After moving to the right by word (4):
| "
    "
| <div>
|   "hello world"
| "
    "
| <br>
| "
    "
| <div>
|   "red, green,<#selection-caret> & blue."
| "
"

After moving to the right by word (5):
| "
    "
| <div>
|   "hello world"
| "
    "
| <br>
| "
    "
| <div>
|   "red, green, &<#selection-caret> blue."
| "
"

After moving to the right by word (6):
| "
    "
| <div>
|   "hello world"
| "
    "
| <br>
| "
    "
| <div>
|   "red, green, & blue.<#selection-caret>"
| "
"

Before moving backwards:
| "
    "
| <div>
|   "hello world"
| "
    "
| <br>
| "
    "
| <div>
|   "red, green, & blue.<#selection-caret>"
| "
"

After moving to the left by word (1):
| "
    "
| <div>
|   "hello world"
| "
    "
| <br>
| "
    "
| <div>
|   "red, green, & <#selection-caret>blue."
| "
"

After moving to the left by word (2):
| "
    "
| <div>
|   "hello world"
| "
    "
| <br>
| "
    "
| <div>
|   "red, green, <#selection-caret>& blue."
| "
"

After moving to the left by word (3):
| "
    "
| <div>
|   "hello world"
| "
    "
| <br>
| "
    "
| <div>
|   "red, <#selection-caret>green, & blue."
| "
"

After moving to the left by word (4):
| "
    "
| <div>
|   "hello world"
| "
    "
| <br>
| "
    "
| <div>
|   "<#selection-caret>red, green, & blue."
| "
"

After moving to the left by word (5):
| "
    "
| <div>
|   "hello <#selection-caret>world"
| "
    "
| <br>
| "
    "
| <div>
|   "red, green, & blue."
| "
"

After moving to the left by word (6):
| "
    "
| <div>
|   "<#selection-caret>hello world"
| "
    "
| <br>
| "
    "
| <div>
|   "red, green, & blue."
| "
"

Before extending forwards:
| "
    "
| <div>
|   "<#selection-caret>hello world"
| "
    "
| <br>
| "
    "
| <div>
|   "red, green, & blue."
| "
"

After extending to the right by word (1):
| "
    "
| <div>
|   "<#selection-anchor>hello<#selection-focus> world"
| "
    "
| <br>
| "
    "
| <div>
|   "red, green, & blue."
| "
"

After extending to the right by word (2):
| "
    "
| <div>
|   "<#selection-anchor>hello world<#selection-focus>"
| "
    "
| <br>
| "
    "
| <div>
|   "red, green, & blue."
| "
"

After extending to the right by word (3):
| "
    "
| <div>
|   "<#selection-anchor>hello world"
| "
    "
| <br>
| "
    "
| <div>
|   "red,<#selection-focus> green, & blue."
| "
"

After extending to the right by word (4):
| "
    "
| <div>
|   "<#selection-anchor>hello world"
| "
    "
| <br>
| "
    "
| <div>
|   "red, green,<#selection-focus> & blue."
| "
"

After extending to the right by word (5):
| "
    "
| <div>
|   "<#selection-anchor>hello world"
| "
    "
| <br>
| "
    "
| <div>
|   "red, green, &<#selection-focus> blue."
| "
"

After extending to the right by word (6):
| "
    "
| <div>
|   "<#selection-anchor>hello world"
| "
    "
| <br>
| "
    "
| <div>
|   "red, green, & blue.<#selection-focus>"
| "
"

Before extending backwards:
| "
    "
| <div>
|   "hello world"
| "
    "
| <br>
| "
    "
| <div>
|   "red, green, & blue.<#selection-caret>"
| "
"

After extending to the left by word (1):
| "
    "
| <div>
|   "hello world"
| "
    "
| <br>
| "
    "
| <div>
|   "red, green, & <#selection-focus>blue.<#selection-anchor>"
| "
"

After extending to the left by word (2):
| "
    "
| <div>
|   "hello world"
| "
    "
| <br>
| "
    "
| <div>
|   "red, green, <#selection-focus>& blue.<#selection-anchor>"
| "
"

After extending to the left by word (3):
| "
    "
| <div>
|   "hello world"
| "
    "
| <br>
| "
    "
| <div>
|   "red, <#selection-focus>green, & blue.<#selection-anchor>"
| "
"

After extending to the left by word (4):
| "
    "
| <div>
|   "hello world"
| "
    "
| <br>
| "
    "
| <div>
|   "<#selection-focus>red, green, & blue.<#selection-anchor>"
| "
"

After extending to the left by word (5):
| "
    "
| <div>
|   "hello <#selection-focus>world"
| "
    "
| <br>
| "
    "
| <div>
|   "red, green, & blue.<#selection-anchor>"
| "
"

After extending to the left by word (6):
| "
    "
| <div>
|   "<#selection-focus>hello world"
| "
    "
| <br>
| "
    "
| <div>
|   "red, green, & blue.<#selection-anchor>"
| "
"
